orch/pyproject.toml
[project]
name = "orch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "HassanSuhaib71", email = "hassansuhaib71@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "langchain-google-genai>=2.0.11",
    "langgraph>=0.2.74",
    "python-dotenv>=1.0.1",
]

[project.scripts]
orch = "orch:main"
flow = "orch.orch:main"
eval = "orch.eval:main"
react = "orch.react:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
